Debug glouton hamiltonian

+++ GoogleHashCode19/construct_graph.py
@@ -27,7 +27,7 @@ for i in range(n):
     ids[i] = id
     tags[i] = photo[3:]
 
-graph = np.zeros((n,n))
+graph = np.zeros((n,n), dtype = int)
 for i in range(n):
     for j in range(i+1, n):
         interest = compute_interest(tags[i], tags[j])
